Fix for packaging

+++ Prototype/Source/Prototype/Private/CellularAutomata.cpp
@@ -28,9 +28,9 @@ CellularAutomata::CellularAutomata(int height, int width)
 void CellularAutomata::Generate()
 {
 	
-	for (size_t i = 0; i < floorHeight; ++i)
+	for (size_t i = 0; (int)i < floorHeight; ++i)
 	{
-		for (size_t j = 0; j < floorWidth; ++j)
+		for (size_t j = 0; (int)j < floorWidth; ++j)
 		{
 			if ((double)rand() / RAND_MAX < birthChance) gol->Summon(i, j); else gol->Kill(i, j);
 			if (i == 0 || j == 0 || i == floorHeight - 1 || j == floorWidth - 1) gol->Summon(i, j);
@@ -164,9 +164,9 @@ void CellularAutomata::ConnectRooms()
 {
 	field = gol->GetFieldCopy();
 	auto rooms = std::vector<vectorOfIndex>();
-	for (size_t i = 1; i < floorHeight - 1; i++)
+	for (size_t i = 1; (int)i < floorHeight - 1; i++)
 	{
-		for (size_t j = 1; j < floorWidth - 1; j++)
+		for (size_t j = 1; (int)j < floorWidth - 1; j++)
 		{
 			if (!field[i][j])
 			{
@@ -282,10 +282,10 @@ void CellularAutomata::Step()
 bool** CellularAutomata::GetFloorMap() const
 {
 	auto exportField = new bool* [floorHeight];
-	for (size_t i = 0; i < floorHeight; i++)
+	for (size_t i = 0; (int)i < floorHeight; i++)
 	{
 		exportField[i] = new bool[floorWidth];
-		for (size_t j = 0; j < floorWidth; j++)
+		for (size_t j = 0; (int)j < floorWidth; j++)
 		{
 			exportField[i][j] = field[i][j];
 		}
